fix(product_service): change getProducts lambda
