--- product_service/lambda/getProducts/errors.js
@@ -1,6 +1,0 @@
-export class NotFoundError extends Error {
-  constructor(message) {
-    super(message);
-    this.name = "NotFoundError";
-  }
-}
