--- product_service/lambda/getProducts/index.js
@@ -1,3 +1,0 @@
-import { getProducts } from "./handler.js";
-
-export { getProducts };
