+++ product_service/lambda/getProducts/index.mjs
@@ -1,0 +1,3 @@
+import { getProducts } from "./handler.mjs";
+
+export { getProducts };
