+++ product_service/lambda/getProducts/package.json
@@ -1,3 +1,5 @@
 {
+  "name": "get-products-all",
+  "version": "1.0.0",
   "type": "module"
 }
